Ejercicio 1 v3
Flotantes

+++ Ejercicio_0001.cpp
@@ -2,19 +2,19 @@
 
 void  main()
 {
-	int base;
-	int altura;
-	int area;
+	float base;
+	float altura;
+	float area;
 	//Entrada o definicion
 	//Aqui va la carta de amor desesperado(comentario unilinea)
 	printf("Base: ");
-	scanf_s("%i", &base);//Indispensable el '&' 
+	scanf_s("%f", &base);//Indispensable el '&' 
 	printf("Altura: ");
-	scanf_s("%i", &altura);
+	scanf_s("%f", &altura);
 	//Proceso
 	area = base * altura;
 	//Salida
-	printf("Area del rectangulo de %i x %i = %i", base,altura,area);
+	printf("Area del rectangulo de %f x %f = %f", base,altura,area);
 	/*Comentario
 	largo de
 	varias lineas*/
